Move CLI stats structs to their own module
Upcoming feature changes will grow these structs.
Moving them preemptively to keep the feature PR small.
